Added Meziantou.Analyzer and addressed the violations
