+++ Src/FluentAssertions/Equivalency/EquivalencyResult.cs
@@ -1,0 +1,7 @@
+namespace FluentAssertions.Equivalency;
+
+public enum EquivalencyResult
+{
+    ContinueWithNext,
+    AssertionCompleted
+}
