+++ Src/FluentAssertions/Equivalency/GetSubjectId.cs
@@ -1,0 +1,6 @@
+namespace FluentAssertions.Equivalency;
+
+/// <summary>
+/// Allows deferred fetching of the subject ID.
+/// </summary>
+public delegate string GetSubjectId();
