+++ Src/FluentAssertions/Equivalency/INode.cs
@@ -80,8 +80,3 @@ public interface INode
     /// </summary>
     bool RootIsCollection { get; }
 }
-
-/// <summary>
-/// Allows deferred fetching of the subject ID.
-/// </summary>
-public delegate string GetSubjectId();
